refactor(settings activity): abstract out mvez
