+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,7 @@
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.HOME" />
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.LAUNCHER" />
@@ -26,12 +27,20 @@
         <activity
             android:name=".SettingsActivity"
             android:label="@string/title_activity_settings"
-            android:screenOrientation="portrait"
             android:parentActivityName=".SearchActivity"
+            android:screenOrientation="portrait"
             android:theme="@android:style/Theme.Material.Light">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.kiyui.timur.mvez.SearchActivity" />
+        </activity>
+        <activity
+            android:name=".MVEZActivity"
+            android:label="@string/title_activity_mvez"
+            android:parentActivityName=".SettingsActivity"
+            android:screenOrientation="portrait"
+            android:theme="@android:style/Theme.Material.Light">
+
         </activity>
     </application>
