+++ app/src/main/res/layout/activity_settings.xml
@@ -51,49 +51,36 @@
         </LinearLayout>
     </LinearLayout>
 
-    <Spinner
-        android:id="@+id/searchSpinner"
-        style="@android:style/Widget.Material.Spinner.Underlined"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/scrollSettings" />
+    <LinearLayout
+        android:id="@+id/mvezSettings"
+        android:layout_width="match_parent"
+        android:layout_height="56dp"
+        android:layout_marginTop="16dp"
+        app:layout_constraintTop_toBottomOf="@id/scrollSettings">
 
-    <EditText
-        android:id="@+id/bangValue"
-        style="@android:style/TextAppearance.Material.Widget.EditText"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="8dp"
-        android:layout_marginTop="8dp"
-        android:ems="10"
-        android:hint="Shortcut"
-        android:inputType="textPersonName"
-        android:paddingLeft="12dp"
-        android:textColor="@color/hint_grey"
-        android:textColorHint="@color/text_hint_grey"
-        app:layout_constraintEnd_toStartOf="@+id/bangSubmit"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/searchSpinner" />
+        <RelativeLayout
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="1"
+            android:gravity="center_vertical"
+            android:orientation="vertical"
+            android:paddingLeft="12dp">
 
-    <Button
-        android:id="@+id/bangSubmit"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Create"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/searchSpinner" />
+            <TextView
+                android:id="@+id/mvezTitle"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="MVEZ Shortcuts"
+                android:textColor="?android:attr/textColorPrimary"
+                android:textSize="16dp" />
 
-    <ListView
-        android:id="@+id/bangList"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginTop="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/bangSubmit" />
+            <TextView
+                style="@android:style/Widget.Material.TextView"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/mvezTitle"
+                android:text="Configure bang syntax shortcuts" />
+        </RelativeLayout>
+    </LinearLayout>
 
 </android.support.constraint.ConstraintLayout>
