+++ app/src/main/res/values/strings.xml
@@ -3,7 +3,7 @@
     <string name="app_name">MVEZ</string>
     <string name="search_view_hint">Type something…</string>
     <string name="title_activity_settings">Settings</string>
-    <string name="title_activity_settings2">Settings</string>
+    <string name="title_activity_mvez">MVEZ Settings</string>
 
     <!-- Strings related to Settings -->
